refactor: split models into individual files
